feature: update kubernetes resource event source
